requirements.txt
amqp==2.3.2
billiard==3.5.0.5
captcha==0.3
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
configparser==3.7.4
cryptowatch-client==0.1.4
Django==2.2.1
django-braces==1.13.0
django-celery-results==1.0.4
django-countries==5.3.3
django-model-utils==3.1.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
django-star-ratings==0.7.0
django-widget-tweaks==1.4.3
gnupg==2.3.1
idna==2.8
kombu==4.2.2.post1
mysqlclient==1.4.2.post1
Pillow==6.0.0
psutil==5.6.2
PyMySQL==0.9.3
python-bitcoinrpc==1.0
pytz==2018.9
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
swapper==1.1.0
urllib3==1.25.2
vine==1.2.0
